requirements.txt
boto3==1.34.69
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9 